FFS-2988: show feedback link in help modal for non-caseworkers (#1036)
Previously, the "Report feedback or bugs" button in the help modal only showed if you had a caseworker-facing feedback form set up in our client agency config. But we want this button to link to our normal CTA that appears in the footer when the user is an applicant. Now, when you're not a caseworker, the button shows and links to that form.